+++ clients/python/balancing_services/models/balancing_capacity_volume.py
@@ -1,10 +1,11 @@
 from __future__ import annotations
 
 from collections.abc import Mapping
-from typing import TYPE_CHECKING, Any, TypeVar
+from typing import Any, TypeVar, TYPE_CHECKING
 
 from attrs import define as _attrs_define
 from attrs import field as _attrs_field
+
 
 if TYPE_CHECKING:
     from ..models.period import Period